CC-36 Structure refactoring
