+++ src/components/Content/Home/NewsBlock/NewsCard/NewsCard.tsx
