--- src/components/Menu/MenuItemDropdown/MenuItemDropdown.module.css
@@ -1,12 +1,0 @@
-.wrapper {
-    max-width: 600px;
-    height: 300px;
-    background-color: hsl(209, 88%, 24%);
-}
-
-.heading {
-    color: #fff;
-    font-weight: bold;
-    padding: 20px;
-    font-family: "Helvetica", serif;
-}
